Conversion of simulation complete
The drift chamber simulation has now been completely converted from the old code. Although minor changes might be made, for now it is functional. When run from drift_diffusion.py, a randomly generated cosmic ray is simulated in the chamber and an animation produced. This is then saved as a gif and shown to the user.